pyproject.toml
[project]
name = "scrape-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
